+++ src/solver/traits.rs
@@ -16,6 +16,7 @@
 
 use crate::physics::traits::PhysicalState;
 use crate::solver::scenario::Scenario;
+use serde::{Deserialize, Serialize};
 use std::{collections::HashMap, string::ToString};
 
 // ============================================================================
@@ -58,7 +59,7 @@ use std::{collections::HashMap, string::ToString};
 /// };
 /// ```
 
-#[derive(Clone, Debug, PartialEq)]
+#[derive(Clone, Debug, PartialEq, Deserialize, Serialize)]
 pub enum SolverType {
     /// Time evolution solution (ODE/PDE integration)
     ///
@@ -254,7 +255,7 @@ impl SolverType {
 ///     }
 /// );
 /// ```
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct SolverConfiguration {
     /// Type of solver and its paraméters
     pub solver_type: SolverType,
@@ -403,7 +404,7 @@ impl SolverConfiguration {
 ///     final_state,
 /// );
 /// ```
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct SimulationResult {
     // Time points at which the solution was computed
     ///
@@ -535,7 +536,10 @@ impl SimulationResult {
 /// use chrom_rs::solver::{Solver, SolverConfiguration, Scenario, DomainBoundaries, RK4Solver};
 /// # use chrom_rs::physics::{PhysicalModel, PhysicalState, PhysicalQuantity, PhysicalData};
 /// # use nalgebra::DVector;
+/// # use serde::{Deserialize, Serialize};
+/// # #[derive(Deserialize, Serialize)]
 /// # struct MyModel;
+/// # #[typetag::serde]
 /// # impl PhysicalModel for MyModel {
 /// #     fn points(&self) -> usize { 1 }
 /// #     fn compute_physics(&self, state: &PhysicalState) -> PhysicalState { state.clone() }